+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [Branch Update: feature/minor-polish] - 2026-04-28
+
+### Changed
+- Branch update commit for `feature/minor-polish` to capture current in-progress module changes.
+
 ## [0.2.7] - 2026-04-25
 
 ### Added